Add AddCommand classes
